Split the classes in `ClaimsAndChecks.kt` off to respective files.
