+++ javatests/arcs/core/data/InformationFlowLabelTest.kt
@@ -18,7 +18,7 @@ import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
 
 @RunWith(JUnit4::class)
-class ClaimsAndChecksTest {
+class InformationFlowLabelTest {
     @Test
     fun prettyPrintSemanticTag() {
         assertThat("${SemanticTag("packageName")}").isEqualTo("packageName")
